Replaced double with float, because we don't need that level of accuracy bruh

+++ include/math/Matrix3.hpp
@@ -7,12 +7,12 @@ struct Matrix3
 {
    // CONSTRUCTORS
    Matrix3();
-   Matrix3(double in);
+   Matrix3(float in);
    Matrix3(Vector3 in);
    Matrix3(Vector3 row1, Vector3 row2, Vector3 row3);
-   Matrix3(double in1, double in2, double in3,
-           double in4, double in5, double in6,
-           double in7, double in8, double in9);
+   Matrix3(float in1, float in2, float in3,
+           float in4, float in5, float in6,
+           float in7, float in8, float in9);
    Matrix3(const Matrix3& other);
 
    ~Matrix3();
@@ -28,17 +28,17 @@ struct Matrix3
    Matrix3 &operator=(const Matrix3& other);
 
    Matrix3 operator+(const Matrix3& other);
-   Matrix3 operator+(double a);
+   Matrix3 operator+(float a);
 
    Matrix3 operator-(const Matrix3& other);
-   Matrix3 operator-(double a);
+   Matrix3 operator-(float a);
 
    Matrix3 operator*(const Matrix3& other);
    Vector3 operator*(const Vector3& other);
-   Matrix3 operator*(double a);
+   Matrix3 operator*(float a);
 
    Matrix3 operator/(const Matrix3& other);
-   Matrix3 operator/(double a);
+   Matrix3 operator/(float a);
 
    // DATA
    Vector3 row1;
